dev-requirements.txt
pylint
pylint-pydantic
flake8
python-lsp-server
